+++ overlay-react/src/layout/index.ts
@@ -6,7 +6,6 @@
  * @module layout
  */
 
-export { computeTooltipPosition } from './computeTooltipPosition';
 export { computeQuadrantPosition } from './computeQuadrantPosition';
 // TODO: Export computeSpotlightPosition when implemented
 // export { computeSpotlightPosition } from './computeSpotlightPosition';